Setup a new ITweaker for responding to the Hypixel Mod API. Currently the API isnt used so the tweaker being used either. The tweaker: - calls essential tweaker to start essential - searches mods in mods folders - searches mcmod.info files in mods - halts the game if mod api is not found This halt should be accompanied with a prompt telling the user how to get the Mod API mod before quitting the game.
When the tweaker should be used, replace "gg.essential.loader.stage0.EssentialSetupTweaker" with "at.yedel.yedelmod.loader.YedelModTweaker" in build.gradle.kts.

It would be nice to use an IFMLLoadingPlugin instead of replacing the essential tweaker, but trying to do so didn't work.

Also add a warning to random placeholders in the config