Implement the Hypixel Mod API: - Added new ping method for Hypixel's ping packet - Location should be much easier to get - Limbo check should be successful now and harder to bypass Privated constructors of singleton classes Updated TextUtils functions Seperated tests and added some more UpdateManager now uses Chat.display Updated AutoGuildWelcome to use regex Add YedelMod as the game profile
With this update, you will need to download the Hypixel Mod API mod and place it into run/mods when running the mod in dev.

This commit was supposed to be the major commit to release the next version for 100 commits, but there are still many things that need to be changed so it's likely it won't happen.